Added support for partial regions

+++ apps/vaporgui/params/histo.h
@@ -33,7 +33,7 @@ public:
 	Histo(int numberBins, float mnData, float mxData);
 	//Special constructor for unsigned char data:
 	//
-	Histo(unsigned char* data, int min_dim[3], int max_dim[3], 
+	Histo(unsigned char* data, size_t min_dim[3], size_t max_dim[3], 
 			 size_t min_bdim[3],size_t max_bdim[3], float mnData, float mxData);
 	~Histo();
 	void reset(int newNumBins = -1);
